refactor: Consolidate Agoda test cases into single AgodaHotelTest class
- Merged TC01_SearchAndSortHotelTest and TC02_SearchAndFilterHotelTest into AgodaHotelTest.java
- Consolidated shared verification methods to reduce code duplication:
  * verifyHomepageIsDisplayed()
  * verifySearchResults()
  * verifyPriceSorting()
  * verifyDestinationInResults()
  * verifyFiltersApplied()
  * verifyPriceFilterRemoved()
- Renamed test methods to tc01_testSearchAndSortHotel() and tc02_testSearchAndFilterHotel()
- Added placeholder tc03_testFutureScenario() for future test case implementation
- Improved code maintainability and reusability across all Agoda hotel test scenarios
- All test cases validated and passing with optimized shared verification logic

Benefits:
- Single test class for all Agoda hotel booking scenarios
- Eliminated redundant verification code between test cases
- Improved scalability for future test case additions
- Enhanced code organization and maintainability
